% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{AddBinObjects}
\alias{AddBinObjects}
\title{A small function to add the objects needed for the binning QP to the problem list.}
\usage{
AddBinObjects(P, s)
}
\arguments{
\item{P}{A list with the problem details.}

\item{s}{A vector of centers.}
}
\value{
A modified problem list.
}
\description{
A small function to add the objects needed for the binning QP to the problem list.
}
\keyword{internal}
