% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{BinningStep}
\alias{BinningStep}
\title{Carry out the binning step.}
\usage{
BinningStep(P)
}
\arguments{
\item{P}{The list of problem objects.}
}
\value{
The input list, with extra members modified.
}
\description{
Carry out the binning step.
}
\keyword{internal}
