% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{BuildConCheckGrid}
\alias{BuildConCheckGrid}
\title{A function to add the constraint-checking grid}
\usage{
BuildConCheckGrid(P)
}
\arguments{
\item{P}{List with problem details.}
}
\value{
Modified list.
}
\description{
A function to add the constraint-checking grid
}
\keyword{internal}
