% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{BuildConstraints}
\alias{BuildConstraints}
\title{Build objects Apos, Aunit, Aeq, Ashape, bshape}
\usage{
BuildConstraints(P)
}
\arguments{
\item{P}{A list of problem obejcts}
}
\value{
The same list, with Apos, Aunit, Aeq, Ashape, bshape filled in.
}
\description{
This function builds the matrices/vectors needed to implement shape constraints in the
estimation step.
}
\keyword{internal}
