% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CleanWeights}
\alias{CleanWeights}
\title{Clean out near-zeros from a probability weights vector and re-normalize}
\usage{
CleanWeights(w)
}
\arguments{
\item{w}{A vector of probability weights}
}
\value{
A vector with the near-zeros removed, still summing to 1.
}
\description{
Clean out near-zeros from a probability weights vector and re-normalize
}
\keyword{internal}
