% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{displayFailure}
\alias{displayFailure}
\title{Display estimation results to console in an unsuccessful case.}
\usage{
displayFailure(pts)
}
\arguments{
\item{pts}{The important points.}
}
\description{
Display estimation results to console in an unsuccessful case.
}
\keyword{internal}
