% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{displaySuccess}
\alias{displaySuccess}
\title{Display estimation results to console in a successful case.}
\usage{
displaySuccess(pts, value)
}
\arguments{
\item{pts}{The important points.}

\item{value}{The objective function value.}
}
\description{
Display estimation results to console in a successful case.
}
\keyword{internal}
