% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{EstimationStep}
\alias{EstimationStep}
\title{Carry out the shape-constrained estimation}
\usage{
EstimationStep(P)
}
\arguments{
\item{P}{The problem list}
}
\value{
Problem list with estimate objects.
}
\description{
Carry out the shape-constrained estimation
}
\keyword{internal}
