% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getQuantile}
\alias{getQuantile}
\title{Estimate a specific quantile of a pdf given abscissa and ordinate values.}
\usage{
getQuantile(x, y, p)
}
\arguments{
\item{x}{The abscissa values of the pdf}

\item{y}{The ordinate values of the pdf}

\item{p}{The probability at which to evaluate the quantile.}
}
\value{
The estimated quantile.
}
\description{
Estimate a specific quantile of a pdf given abscissa and ordinate values.
}
\keyword{internal}
