% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getQuantileFunction}
\alias{getQuantileFunction}
\title{Build a quantile function for a given constrained density estimator.}
\usage{
getQuantileFunction(x)
}
\arguments{
\item{x}{An object of S3 class \code{scdensity}.}
}
\value{
A function that takes a fraction and returns the quantile.
}
\description{
This function implements a crude but numerically reliable method to approximate the
quantile function of an \code{scdensity} estimate.
}
\keyword{internal}
