% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{InitializeP}
\alias{InitializeP}
\title{Initialize the list of problem-related objects.}
\usage{
InitializeP(x, h, constraints, method, opts)
}
\arguments{
\item{x}{The data vector.}

\item{h}{The bandwidth (a positive scalar).}

\item{constraints}{The vector of constraint strings.}

\item{method}{Either "weightedKDE" or "adjustedKDE".}
}
\value{
A list in the form needed by WeightedKDE(), with additional elements initialized
to their default values.
}
\description{
Initialize the list of problem-related objects.
}
\keyword{internal}
