% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpenedKDE.R
\name{isBoundedR}
\alias{isBoundedR}
\title{Check for zeros at the right side of a vector of function values.}
\usage{
isBoundedR(f, ix)
}
\arguments{
\item{f}{A vector of function values for increasing abscissa values.}

\item{ix}{An index giving the cutoff for checking for zero.}
}
\value{
A logical value indicating if the constraint is satisfied.
}
\description{
Given a vector of n function values and an index ix, check whether values with indices
greater than or equal to ix are zero. Returns TRUE if they are, FALSE otherwise.
}
\details{
As in \code{isUnimodal}, the values are first scaled to fill [0, 1] and then rounded to
four decimal places.  Because of this it is still possible to use the "bounded support"
constraints with the Gaussian kernel.

This function is intended to be called from other functions in the scdensity package.
It does not implement any argument checking.
}
\keyword{internal}
