% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpenedKDE.R
\name{isUnimodal}
\alias{isUnimodal}
\title{Check for unimodality of function values.}
\usage{
isUnimodal(f)
}
\arguments{
\item{f}{A vector of function values for increasing abscissa values.}
}
\value{
A logical value indicating if unimodality is satisfied.
}
\description{
Given a set of function values for increasing abscissa values, we call this unimodal if
there are zero or one values that are greater than all of their neighbors. Before
checking for modes, the values are scaled to fill [0, 1] and then rounded to four
decimal places.  This eliminates unwanted detection of tiny differences as modes.
}
\details{
This function is intended to be called from other functions in the scdensity package.
It does not implement any argument checking.
}
\keyword{internal}
