% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{MakeCenters}
\alias{MakeCenters}
\title{Create a vector of kernel centers covering [LB, UB].}
\usage{
MakeCenters(LB, UB, width, minspace, PoS = NULL)
}
\arguments{
\item{LB}{The lower bound.}

\item{UB}{The upper bound.}

\item{width}{The maximum possible spacing.}

\item{minspace}{The minimum allowable spacing.}

\item{PoS}{Point of symmetry (default NULL)}
}
\value{
A vector of kernel centers.
}
\description{
Make the spacing as large as possible without going over \code{width}.
If symmetric, ensure we have an even number of centers that are symmetric around \code{PoS}.
}
\keyword{internal}
