% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{makeOF}
\alias{makeOF}
\title{A function factory for making the search objective function.}
\usage{
makeOF(P)
}
\arguments{
\item{P}{The list of problem details.}
}
\value{
The objective function.
}
\description{
Used when we need to search for important points.
P is the problem list. It should have already gone through BuildConCheckGrid and
BinningStep. The returned function must return a value even if WeightedKDE() fails.
In case of failure, just assign a large random value to the objective value (to keep
the search from stagnating or moving systematically in one direction).
}
\keyword{internal}
