% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdensity.R
\name{print.scdensity}
\alias{print.scdensity}
\title{Print method for class \code{scdensity}.}
\usage{
\method{print}{scdensity}(x, ...)
}
\arguments{
\item{x}{An object of S3 class \code{scensity}.}

\item{...}{Included for consistency with generic functions.}
}
\description{
Displays the names of the elements of the scdensity list object
and their sizes and types. Includes minimal comments about the most important elements.
}
