% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdensity.R
\name{print.summary.scdensity}
\alias{print.summary.scdensity}
\title{Prints the information in a \code{summary.scdensity} object to the console.}
\usage{
\method{print}{summary.scdensity}(x, ...)
}
\arguments{
\item{x}{An object of S3 class \code{summary.scdensity}.}

\item{...}{Included for consistency with generic functions.}
}
\description{
Prints the information in a \code{summary.scdensity} object to the console.
}
