% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdensity.R
\name{summary.scdensity}
\alias{summary.scdensity}
\title{Summary method for class \code{scdensity}.}
\usage{
\method{summary}{scdensity}(object, ...)
}
\arguments{
\item{object}{An object of S3 class \code{scensity}.}

\item{...}{Included for consistency with generic functions.}
}
\description{
Collects high-level information about the \code{scdensity} object and some descriptive
statistics.
}
