% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKDE.R
\name{WeightedKDE}
\alias{WeightedKDE}
\title{Function to carry out the weighted or adjusted KDE optimization.}
\usage{
WeightedKDE(P)
}
\arguments{
\item{P}{A list, as created by InitializeP().}
}
\value{
A mutated version of the input list, with additional elements giving the optimization
        output.
}
\description{
This function sets up the problem and finds an optimal shape-constrained estimate for a
specified set of important points.
}
\keyword{internal}
