% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdensity-package.R
\docType{package}
\name{scdensity-package}
\alias{scdensity-package}
\title{scdensity: Shape-constrained kernel density estimation.}
\description{
This package computes one-dimensional Gaussian kernel density estimates subject to a
variety of shape constraints, including unimodality, bimodality, symmetry, and others.
}
\details{
All of the package's functionality can be accessed through the function
\code{\link[scdensity]{scdensity}}.  See that function's help file.  The function is used
just like \code{\link[stats]{density}}, but with extra arguments for handling the constraints.
}
